pyproject.toml
[project]
name = "ai-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flet[all]==0.28.3",
    "google-generativeai>=0.8.5",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
]
